[mango_autoencoder] Migration, UV Integration, and Docs/CI Updates (#338)
* Add initial project structure with pyproject.toml, README, and TODO files

- Created pyproject.toml for project configuration and dependencies.
- Added README.md for project documentation.
- Included TODO.md to outline future tasks.
- Implemented initial source files: __init__.py, calendar_features.py, date_utils.py for core functionality.
- Established uv.lock for dependency management.

* Add workflow for publishing mango-calendar to PyPI and include LICENSE file

- Created a GitHub Actions workflow to automate the publishing of the mango-calendar package to PyPI upon tagging.
- Added LICENSE file with Apache License 2.0 for project compliance.
- Updated pyproject.toml with project metadata and dependencies.
- Enhanced README.md with detailed project features and installation instructions.
- Implemented run_tests.py for unit testing execution.
- Removed TODO.md as tasks are now addressed in the project structure.

* Missing typehint and fix a line too long

* Removed get_covid_lockdown() func

* Removed unused imports

* Added test workflow for mango_calendar, restored function get_covid_lockdowns() and deleted function get_mwc()

* Removed hello function

* Removed test for hello function

* Removed run_test scripts and translated pyproject comments to english

* Added compatibilty with 3.10, 3.11 and 3.13

* Removed mentions to get_mwc in the readme file

* Mango calendar test action now ignores changes in .md files

* path-ignore does not work when i have a paths in the same section

* Update PyPI workflow to specify package name and add versioning to __init__.py

* Update PyPI workflow to trigger on release events and adjust token usage for publishing

* Refactor PyPI workflow to use relative paths for package build and publish steps

* Update pyproject.toml to use setuptools for build system and dynamic versioning

* Add __init__.py file and correct description in pyproject.toml

* Add initial __init__.py for Mango Calendar package and define public API

* Update workflow to trigger on tag pushes for Mango Calendar publication

* Initialize mango_autoencoder package and move models/ from mango_time_series #324

* Add initial GitHub Actions workflow files for CI #331

* Structuring autoencoder documentation

* remove tensor flow from requirements.txt in mango_time_series

* adding a custom autoencoder logger

* fix imports

* reestructring dependecies and the use of uv

* delete file

* add docs files

* delete files when testing

* updating pyproject.toml

* add README.md content

* optimize imports

* update uv.lock

* add auto_doc monck imports

* correction in utils.rst

* add mango calendar docs

* delete mango calendar files in mango_time_series

* add the versions to the documentation libraries

* set dependency on the tensorflow library

* set dependency on the tensorflow library

* delete python 3.13 in autoencoder workflow

* set dependency on the tensorflow library

* set up python version to 3.11

* update pyproject toml in autoencoder

* update pyproject toml in autoencoder

* remove tensorflow-io version

* handle tensorflow-io version

* handle tensorflow-io version

* remove macos

* delete decodoc

* move tests

* update README.md

* add working directory in unit testing workflow

* add working directory in unit testing workflow

* change path in testing uit workflows

---------

+++ mango_autoencoder/mango_autoencoder/tests/test_autoencoder.py
@@ -1,19 +1,17 @@
 # isort:skip_file
-import tensorflow as tf
-
 import shutil
-import tempfile
 import unittest
 from pathlib import Path
 
 import numpy as np
 import pandas as pd
 import polars as pl
+import tensorflow as tf
 from keras.src.layers import Bidirectional
 
-from mango_time_series.models.utils import processing
-from mango_time_series.models.autoencoder import AutoEncoder
-from mango_time_series.models.modules import decoder, encoder
+from mango_autoencoder.autoencoder import AutoEncoder
+from mango_autoencoder.modules import decoder, encoder
+from mango_autoencoder.utils import processing
 
 
 class TestEncoder(unittest.TestCase):
@@ -463,7 +461,8 @@ class TestAutoEncoderCases(unittest.TestCase):
         :return: None
         :rtype: None
         """
-        self.base_dir = Path(tempfile.mkdtemp())
+        self.root_dir = Path.cwd()
+        self.base_dir = self.root_dir / "test_autoencoder_cases"
         self.test_cases = [
             {
                 "with_ids": False,
@@ -509,11 +508,17 @@ class TestAutoEncoderCases(unittest.TestCase):
 
         Removes the temporary directory and all its contents to ensure
         a clean state for the next test.
-
-        :return: None
-        :rtype: None
         """
-        shutil.rmtree(self.base_dir)
+
+        autoencoder_dir = self.root_dir / "autoencoder"
+        plots_dir = self.root_dir / "plots"
+        if autoencoder_dir.exists() and autoencoder_dir.is_dir():
+            shutil.rmtree(autoencoder_dir)
+
+        if plots_dir.exists() and plots_dir.is_dir():
+            shutil.rmtree(plots_dir)
+        if self.base_dir.exists() and self.base_dir.is_dir():
+            shutil.rmtree(self.base_dir)
 
     @staticmethod
     def generate_synthetic_data_standard(num_samples=500, num_features=3, **kwargs):
